Remove outdated cache file
